factored out the reading of generated files
